paths rendered on map on load
