+++ src/components/App.jsx
@@ -9,7 +9,6 @@ import CateringForm from './CateringForm';
 import { AuthProvider } from '../contexts/authContext';
 import Error404 from '../pages/Error404';
 import Map from './Map';
-// import StaticLocations from './StaticLocations';
 
 
 function App() {
